requirements.txt
python-telegram-bot==21.5
mistralai==1.2.5
python-dotenv==1.0.0
Pillow==10.2.0